+++ src/app/AdminComponents/admin-cruds/admin-cancha-details/admin-cancha-details.ts
@@ -86,6 +86,8 @@ export class AdminCanchaDetails {
   }
 
   handleDelete() {
+    if (!confirm(`¿Seguro que quiere eliminar esta cancha?`)) return;
+
     this.canchaService.deleteCancha(this.cancha()?.id!).subscribe({
         next: (res) => {
           console.log('✅ Cancha eliminada con éxito:', res);
